Fix bug where trailing slash would be added to file links
Previously it was possible to end on a version home page without a
trailing slash. If you then clicked on detail page, it would clobber the
version number and you would get a 404. To fix this we tried to add a
trailing slash if you visited a page without one. This caused issues
when you visited a page that should not end in a trailing slash (*.html
for example).

Now links to the detail pages will include the version number to prevent
this issue.